+++ limberframework/authentication/authentication_service_provider.py
@@ -12,7 +12,8 @@ class AuthServiceProvider(ServiceProvider):
     """Registers database services to the service container."""
 
     def register(self) -> None:
-        def register_authenticator(app: Application):
-            return make_authenticator(app["config"]["auth"])
+        async def register_authenticator(app: Application):
+            config_service = await app.make("config")
+            return await make_authenticator(config_service["auth"])
 
         self.app.bind("auth", register_authenticator, defer=True)