+++ tests/database/test_database_middleware.py
@@ -1,4 +1,4 @@
-from unittest.mock import AsyncMock, MagicMock, Mock
+from unittest.mock import AsyncMock, Mock
 
 from pytest import mark, raises
 
@@ -7,22 +7,22 @@ from limberframework.database.middleware import DatabaseSessionMiddleware
 
 @mark.asyncio
 async def test_dispatch_database_session_middleware_without_exception():
-    mock_request = MagicMock()
-    mock_request.app.__getitem__.return_value = Mock()
+    mock_request = Mock()
+    mock_request.app.make = AsyncMock(return_value=Mock())
 
     mock_call_next = AsyncMock()
 
     middleware = DatabaseSessionMiddleware(Mock())
     await middleware.dispatch(mock_request, mock_call_next)
 
-    mock_request.app.__getitem__.return_value.commit.assert_called_once()
-    mock_request.app.__getitem__.return_value.close.assert_called_once()
+    mock_request.state.db.commit.assert_called_once()
+    mock_request.state.db.close.assert_called_once()
 
 
 @mark.asyncio
 async def test_dispatch_database_session_middleware_with_exception():
-    mock_request = MagicMock()
-    mock_request.app.__getitem__.return_value = Mock()
+    mock_request = Mock()
+    mock_request.app.make = AsyncMock(return_value=Mock())
 
     mock_call_next = AsyncMock()
     mock_call_next.side_effect = Exception()
@@ -32,5 +32,5 @@ async def test_dispatch_database_session_middleware_with_exception():
     with raises(Exception):
         await middleware.dispatch(mock_request, mock_call_next)
 
-    mock_request.app.__getitem__.return_value.commit.assert_not_called()
-    mock_request.app.__getitem__.return_value.close.assert_called_once()
+    mock_request.state.db.commit.assert_not_called()
+    mock_request.state.db.close.assert_called_once()