+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - test_make_store_file_store test to pass cache path inside the config dictionary.
 - circular import in support/service_provider between ServiceProvider and Application.
+- test_async_redis_add_key_does_not_exist to mock the expireat method with AsyncMock.
+- storing locks in a Locker by correcting the `_locks` attribute name in `AsyncRedisLocker`.
+- the passing of cache configuration settings from `make_locker()` to `AsyncRedisLocker`.
+- the setting of an expire time for a resource in a Redis cache using `AsyncRedisStore`.
+- issues with mocking in test for checking the prevention of too many requests, `test_dispatch_too_many_requests_exception`.
+- `make_locker` raising `ValueError` for unknown locker.
 
 ### Added
 - deferrable services which will be loaded when needed.
@@ -18,6 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - pytest pre-commit hook to run tests before committing.
 - tests to authentication, database middleware, routing middleware, and service providers to increase test coverage to 100%.
 - isort pre-commit hook for consistent ordering of imports.
+- support for async services by Application.
+- AsyncRedisStore to support asynchronous Redis communication.
+- cache locker service to lock resources in the cache database, including the AsyncRedisLocker service.
+- tests for `locker` module.
 
 ### Changed
 - restructured tests to match project structure for consistency.